+++ app/components/article/ArticleForm.vue
@@ -1,0 +1,143 @@
+<script setup lang="ts">
+import { computed, ref } from 'vue'
+import { useForm } from 'vee-validate'
+import { Loader2 } from 'lucide-vue-next'
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
+import TinyMceEditor from './TinyMceEditor.vue'
+import PublishOptions from './PublishOptions.vue'
+import { articleFormSchema } from '@/composables/useArticleForm'
+import type { ArticleFormValues } from '@/composables/useArticleForm'
+
+interface Props {
+  initialValues?: Partial<ArticleFormValues>
+  loading?: boolean
+}
+
+const props = withDefaults(defineProps<Props>(), {
+  initialValues: () => ({}),
+  loading: false
+})
+
+const emit = defineEmits<{
+  saveDraft: [values: ArticleFormValues]
+  publish: [values: ArticleFormValues]
+}>()
+
+// Form setup
+const form = useForm({
+  validationSchema: articleFormSchema,
+  initialValues: {
+    title: props.initialValues?.title || '',
+    slug: props.initialValues?.slug || '',
+    content: props.initialValues?.content || '',
+    publishMode: (props.initialValues?.publishMode || 'immediate') as 'immediate' | 'schedule',
+    scheduledDateTime: props.initialValues?.scheduledDateTime || ''
+  }
+})
+
+// Track publishMode for conditional rendering
+const currentPublishMode = ref<'immediate' | 'schedule'>(
+  (props.initialValues?.publishMode as 'immediate' | 'schedule') || 'immediate'
+)
+
+// Computed
+const isScheduleMode = computed(() => currentPublishMode.value === 'schedule')
+
+// Submit handlers
+const handleSaveAsDraft = form.handleSubmit((values) => {
+  emit('saveDraft', values)
+})
+
+const handlePublish = form.handleSubmit((values) => {
+  emit('publish', values)
+})
+</script>
+
+<template>
+  <form class="space-y-6">
+    <!-- Title -->
+    <FormField v-slot="{ componentField }" name="title">
+      <FormItem>
+        <FormLabel>Title *</FormLabel>
+        <FormControl>
+          <Input v-bind="componentField" placeholder="Enter article title" :disabled="loading" />
+        </FormControl>
+        <FormMessage />
+      </FormItem>
+    </FormField>
+
+    <!-- Slug -->
+    <FormField v-slot="{ componentField }" name="slug">
+      <FormItem>
+        <FormLabel>Slug *</FormLabel>
+        <FormControl>
+          <Input v-bind="componentField" placeholder="enter-article-slug" :disabled="loading" />
+        </FormControl>
+        <FormMessage />
+      </FormItem>
+    </FormField>
+
+    <!-- Content -->
+    <FormField v-slot="{ value, handleChange }" name="content">
+      <FormItem>
+        <FormLabel>Content *</FormLabel>
+        <FormControl>
+          <TinyMceEditor
+            :model-value="value"
+            :disabled="loading"
+            @update:model-value="handleChange"
+          />
+        </FormControl>
+        <FormMessage />
+      </FormItem>
+    </FormField>
+
+    <!-- Publish Options -->
+    <div>
+      <FormField v-slot="{ value, handleChange }" name="publishMode">
+        <FormItem>
+          <FormLabel>Publish Options</FormLabel>
+          <FormControl>
+            <PublishOptions
+              :publish-mode="value"
+              :scheduled-date-time="form.values.scheduledDateTime"
+              :disabled="loading"
+              @update:publish-mode="
+                (val) => {
+                  handleChange(val)
+                  currentPublishMode = val
+                }
+              "
+              @update:scheduled-date-time="(val) => form.setFieldValue('scheduledDateTime', val)"
+            />
+          </FormControl>
+          <FormMessage />
+        </FormItem>
+      </FormField>
+
+      <FormField v-if="isScheduleMode" v-slot="{ componentField }" name="scheduledDateTime">
+        <FormItem>
+          <FormControl>
+            <input v-bind="componentField" type="hidden" />
+          </FormControl>
+          <FormMessage />
+        </FormItem>
+      </FormField>
+    </div>
+
+    <!-- Action Buttons -->
+    <div class="flex gap-4 pt-4">
+      <Button variant="outline" type="button" :disabled="loading" @click="handleSaveAsDraft">
+        <Loader2 v-if="loading" class="mr-2 h-4 w-4 animate-spin" />
+        Save as Draft
+      </Button>
+
+      <Button type="button" :disabled="loading" @click="handlePublish">
+        <Loader2 v-if="loading" class="mr-2 h-4 w-4 animate-spin" />
+        Publish
+      </Button>
+    </div>
+  </form>
+</template>
